Bibliography.sorting_rules — Constant

const sorting_rules = Dict{Symbol, Vector{Symbol}}(
    :nty  => [:authors;:editors;:title;:date],
    :nyt  => [:authors;:editors;:date;:title]
);

Implemented sorting rules for bibliography entry sorting.

See also sort_bibliography!.

Bibliography.bibtex_to_web — Method

bibtex_to_web(source::String)

Convert a BibTeX file to a web compatible format, specifically for the StaticWebPages.jl package.

Bibliography.export_bibtex — Method

export_bibtex(e::Entry)
export_bibtex(bibliography::DataStructures.OrderedDict{String,Entry})
export_bibtex(target::String, bibliography::DataStructures.OrderedDict{String,Entry})

Export an entry or a bibliography to BibTeX format.

Bibliography.export_web — Method

export_web(bibliography::DataStructures.OrderedDict{String,BibInternal.Entry})

Export a biblography in internal format to the web format of the StaticWebPages.jl pakcage.

Bibliography.import_bibtex — Method

import_bibtex(file::String)

Import a BibTeX file and convert it to the internal bibliography format.

Bibliography.recursive_isless — Function

recursive_isless(a::Entry, b::Entry, fields::Tuple{Symbol},
                 depth::Int = 0)

Helper function for sort_bibliography!.

This function allows recursive checking if a < b with descending importance. The importance set for the comparison is defined by the argument fields. This argument is a tuple consisting of symbols denoting the fields of the data type BibInternal.Entry. The ordering implies the importance.

The depth argument is purely for iterating/recursive purposes.

Bibliography.select — Method

select(
    bibliography::DataStructures.OrderedDict{String,Entry},
    selection::Vector{String};
    complementary::Bool = false
    )

Select a part of a bibliography based on a given selection set of keys. If complementary is true, selection designates which entries will not be kept. By default, complementary is set to false.

Bibliography.sort_bibliography! — Function

sort_bibliography!(
    bibliography::DataStructures.OrderedDict{String,Entry},
    sorting_rule::Symbol = :key
    )

Sorts the bibliography in place.

The sorting order can be set by specifying the sorting_rule. The sorting is implemented via isless() functions. For detailed insight have a look at the isless() implementation of Julia and BibInternal.jl.

Supported symbols for sorting_rule are:

:key (default): sort by bibliography keys e.g. BibTeX keys or :id

the sorting rules defined in sorting_rules

Note:

The sorting is not following explicitly bibliographic alphabetizing conventions. It follows standard comparator behaviour implied by the implemented isless() functions (string comparators).