Bibliography.sorting_rules — Constant

const sorting_rules = Dict{Symbol, Vector{Symbol}}(
    :nty  => [:authors;:editors;:title;:date],
    :nyt  => [:authors;:editors;:date;:title]
);

Implemented sorting rules for bibliography entry sorting.

See also sort_bibliography!.

Bibliography.Publication — Type

Publication

A structure to store all the information necessary to web export.

Bibliography.Publication — Method

Publication(entry)

Construct a Publication (compatible with web export) from an Entry.

Bibliography.access_to_bibtex! — Method

access_to_bibtex!(fields, a)

Transform the how-to-access field to a BibTeX string.

Bibliography.bibtex_to_web — Method

bibtex_to_web(source::String)

Convert a BibTeX file to a web compatible format, specifically for the StaticWebPages.jl package.

Bibliography.date_to_bibtex! — Method

date_to_bibtex!(fields, date)

Convert a date to a BibTeX string.

Bibliography.eprint_to_bibtex! — Method

eprint_to_bibtex!(fields, eprint)

Convert eprint information to a BibTeX string.

Bibliography.export_bibtex — Method

export_bibtex(target, bibliography)

Export a bibliography to BibTeX format.

Bibliography.export_bibtex — Method

export_bibtex(bibliography)

Export a bibliography to a BibTeX string.

Bibliography.export_bibtex — Method

export_bibtex(e::Entry)

Export an Entry to a BibTeX string.

Bibliography.export_web — Method

export_web(bibliography::DataStructures.OrderedDict{String,BibInternal.Entry})

Export a bibliography in internal format to the web format of the StaticWebPages.jl package. Also used by DocumenterCitations.jl.

Bibliography.field_to_bibtex — Method

field_to_bibtex(key, value)

Convert an entry field to BibTeX format.

Bibliography.import_bibtex — Method

import_bibtex(input)

Import a BibTeX file or parse a BibTeX string and convert it to the internal bibliography format.

Bibliography.in_to_bibtex! — Method

in_to_bibtex!(fields::BibInternal.Fields, i::BibInternal.In)

Convert the "published in" information to a BibTeX string.

Bibliography.int_to_spaces — Method

int_to_spaces(n)

Make a string of n spaces.

Bibliography.name_to_string — Method

name_to_string(name::BibInternal.Name)

Convert a name in an Entry to a string.

Bibliography.names_to_strings — Method

names_to_strings(names)

Convert a collection of names to a BibTeX string.

Bibliography.recursive_isless — Function

recursive_isless(a::Entry, b::Entry, fields::Tuple{Symbol},
                 depth::Int = 0)

Helper function for sort_bibliography!.

This function allows recursive checking if a < b with descending importance. The importance set for the comparison is defined by the argument fields. This argument is a tuple consisting of symbols denoting the fields of the data type BibInternal.Entry. The ordering implies the importance.

The depth argument is purely for iterating/recursive purposes.

Bibliography.select — Method

select(
    bibliography::DataStructures.OrderedDict{String,Entry},
    selection::Vector{String};
    complementary::Bool = false
    )

Select a part of a bibliography based on a given selection set of keys. If complementary is true, selection designates which entries will not be kept. By default, complementary is set to false.

Bibliography.sort_bibliography! — Function

sort_bibliography!(
    bibliography::DataStructures.OrderedDict{String,Entry},
    sorting_rule::Symbol = :key
    )

Sorts the bibliography in place.

The sorting order can be set by specifying the sorting_rule. The sorting is implemented via isless() functions. For detailed insight have a look at the isless() implementation of Julia and BibInternal.jl.

Supported symbols for sorting_rule are:

:key (default): sort by bibliography keys e.g. BibTeX keys or :id

the sorting rules defined in sorting_rules

Note:

The sorting is not following explicitly bibliographic alphabetizing conventions. It follows standard comparator behaviour implied by the implemented isless() functions (string comparators).

Bibliography.xcite — Method

xcite(entry)

Format the BibTeX cite output of an Entry for web export.

Bibliography.xfile — Method

xfile(entry)

Format the downloadable path of an Entry file for web export.

Bibliography.xin — Method

xin(entry)

Format the appears-in field of an Entry for web export.

Bibliography.xlabels — Method

xlabels(entry)

Format the labels of an Entry for web export.

Bibliography.xlink — Method

xlink(entry)

Format the download link of an Entry for web export.

Bibliography.xnames — Function

xnames(entry, editors = false; names = :full)

Format the name of an Entry for web export.

Arguments:

entry: an entry

editors: true if the name describes editors

names: :full (full names) or :last (last names + first name abbreviation)

Bibliography.xtitle — Method

xtitle(entry

Format the title of an Entry for web export.

Bibliography.xyear — Method

xyear(entry)

Format the year of an Entry for web export.