Bibliography.bibtex_to_web — Method

bibtex_to_web(source::String)

Convert a BibTeX file to a web compatible format, specifically for the StaticWebPages.jl pakcage.

Bibliography.export_bibtex — Method

export_bibtex(e::Entry)
export_bibtex(bibliography::DataStructures.OrderedDict{String,Entry})
export_bibtex(target::String, bibliography::DataStructures.OrderedDict{String,Entry})

Export an entry or a bibliography to BibTeX format.

Bibliography.export_web — Method

export_web(bibliography::DataStructures.OrderedDict{String,BibInternal.Entry})

Export a biblography in internal format to the web format of the StaticWebPages.jl pakcage.

Bibliography.import_bibtex — Method

import_bibtex(file::String)

Import a BibTeX file and convert it to the internal bibliography format.

Bibliography.select — Method

select(
    bibliography::DataStructures.OrderedDict{String,Entry},
    selection::Vector{String};
    complementary::Bool = false
    )

Select a part of a bibliography based on a given selection set of keys. If complementary is true, selection designates which entries will not be kept. By default, complementary is set to false.