BibInternal.entries — Constant

const entries = [
    :article,
    :book,
    :booklet,
    :inbook,
    :incollection,
    :inproceedings,
    :manual,
    :masterthesis,
    :misc,
    :phdthesis,
    :proceedings,
    :techreport,
    :unpublished,
]

List of possible entries (currently based on bibtex). Keep it sorted for readability.

BibInternal.fields — Constant

const fields = [
    :address,
    :annote,
    :archivePrefix,
    :author,
    :booktitle,
    :chapter,
    :crossref,
    :edition,
    :editor,
    :eprint,
    :howpublished,
    :institution,
    :journal,
    :key,
    :month,
    :note,
    :number,
    :organization,
    :pages,
    :primaryClass,
    :publisher,
    :school,
    :series,
    :title,
    :type,
    :volume,
    :year
]

List of possible fields (currently based on bibtex). Keep it sorted for readability

BibInternal.maxfieldlength — Constant

const maxfieldlength

For output formatting purpose, for instance, export to BibTeX format.

BibInternal.rules — Constant

const rules = Dict([
    "article"       => ["author", "journal", "title", "year"]
    "book"          => [("author", "editor"), "publisher", "title", "year"]
    "booklet"       => ["title"]
    "eprint"        => ["author", "eprint", "title", "year"]
    "inbook"        => [("author", "editor"), ("chapter", "pages"), "publisher", "title", "year"]
    "incollection"  => ["author", "booktitle", "publisher", "title", "year"]
    "inproceedings" => ["author", "booktitle", "title", "year"]
    "manual"        => ["title"]
    "masterthesis"  => ["author", "school", "title", "year"]
    "misc"          => []
    "phdthesis"     => ["author", "school", "title", "year"]
    "proceedings"   => ["title", "year"]
    "techreport"    => ["author", "institution", "title", "year"]
    "unpublished"   => ["author", "note", "title"]
])

List of BibTeX rules bases on the entry type. A field value as a singleton represents a required field. A pair of values represents mutually exclusive required fields.

BibInternal.AbstractEntry — Type

abstract type AbstractEntry end

Abstract Entry supertype.

BibInternal.Access — Type

struct Access
    doi::String
    howpublished::String
    url::String
end

Store the online access of an entry as a String. Handles the fields doi and url and the arXiV entries. For additional fields or entries, please fill an issue or make a pull request.

BibInternal.Access — Method

Access(fields::Fields)

Construct the online access information based on the entry fields.

BibInternal.Date — Type

struct Date
    day::String
    month::String
    year::String
end

Store the date information as day, month, and year.

BibInternal.Date — Method

Date(fields::Fields)

Construct the date information based on the entry fields.

BibInternal.Entry — Type

struct Entry <: AbstractEntry
    access::Access
    authors::Names
    booktitle::String
    date::Date
    editors::Names
    eprint::Eprint
    id::String
    in::In
    fields::Dict{String,String}
    title::String
    type::String
end

Generic Entry type. If some construction rules are required, it should be done beforehand. Check bibtex.jl as the example of rules implementation for BibTeX format.

BibInternal.Entry — Method

Entry(id::String, fields::Fields)

Construct an entry with a unique id and a list of Fields.

BibInternal.Eprint — Type

struct Eprint
    archive_prefix::String
    eprint::String
    primary_class::String
end

Store the information related to arXiv eprint format.

BibInternal.Eprint — Method

Eprint(fields::Fields)

Construct the eprint arXiv information based on the entry fields. Handle old and current arXiv format.

BibInternal.Fields — Type

Fields = Dict{String, String}

Stores the fields (name => value) of an entry.

BibInternal.In — Type

struct In
    address::String
    chapter::String
    edition::String
    institution::String
    journal::String
    number::String
    organization::String
    pages::String
    publisher::String
    school::String
    series::String
    volume::String
end

Store all the information related to how an entry was published.

BibInternal.In — Method

In(fields::Fields)

Construct the information of how an entry was published based on its fields

BibInternal.Name — Method

Name(str::String)

Decompose without ambiguities a name as particle (optional) last, junior (optional), first middle (optional) based on BibTeX possible input. As for BibTeX, the decomposition of a name in the form of first last is also possible, but ambiguities can occur.

BibInternal.arxive_url — Method

arxive_url(fields::Fields)

Make an arxiv url from an eprint entry. Work with both old and current arxiv BibTeX format.

BibInternal.check_entry — Method

check_entry(fields::Fields)

Check the validity of the fields of a BibTeX entry.

BibInternal.erase_spaces — Method

erase_spaces(str::String)

Erase extra spaces, i.e. r"[ ]+", from str and return a new string.

BibInternal.get_delete! — Method

get_delete!(fields::Fields, key::String)

Get a the value of a field and delete it afterward.

BibInternal.make_bibtex_entry — Method

make_bibtex_entry(id::String, fields::Fields)

Make an entry if the entry follows the BibTeX guidelines. Throw an error otherwise.

BibInternal.names — Method

names(str::String)

Decompose into parts a list of names in BibTeX compatible format. That is names sparated by and.

BibInternal.space — Method

space(field::Symbol)

Return the amount of spaces needed to export entries, for instance to BibTeX format.